monte

+++ monte_carlo.c
@@ -20,7 +20,7 @@ float mc_pi(int n);
         ps++;
 	}
 	}
-	float pi = 4.0*pc/(ps+pc);
+	float pi = 4.0* (float)pc/(ps+pc);
 	return pi;
 }
 float frandom();
